--- tests/Gedmo/Tree/Fixture/ForeignUuidCategory.php
@@ -1,133 +1,0 @@
-<?php
-
-namespace Tree\Fixture;
-
-use Gedmo\Mapping\Annotation as Gedmo;
-use Doctrine\ORM\Mapping as ORM;
-use Ramsey\Uuid;
-
-/**
- * @ORM\Entity(repositoryClass="Gedmo\Tree\Entity\Repository\NestedTreeRepository")
- * @Gedmo\Tree(type="nested")
- */
-class ForeignUuidCategory
-{
-    /**
-     * @ORM\Column(name="id", type="uuid_binary")
-     * @ORM\Id
-     */
-    private $id;
-
-    /**
-     * @ORM\Column(name="title", type="string", length=64)
-     */
-    private $title;
-
-    /**
-     * @Gedmo\TreeLeft
-     * @ORM\Column(name="lft", type="integer")
-     */
-    private $lft;
-
-    /**
-     * @Gedmo\TreeRight
-     * @ORM\Column(name="rgt", type="integer")
-     */
-    private $rgt;
-
-    /**
-     * @Gedmo\TreeParent
-     * @ORM\ManyToOne(targetEntity="ForeignUuidCategory", inversedBy="children")
-     * @ORM\JoinColumns({
-     *   @ORM\JoinColumn(name="parent_id", referencedColumnName="id", onDelete="CASCADE")
-     * })
-     */
-    private $parent;
-
-    /**
-     * @Gedmo\TreeRoot
-     * @ORM\ManyToOne(targetEntity="ForeignUuidCategory")
-     */
-    private $root;
-
-    /**
-     * @Gedmo\TreeLevel
-     * @ORM\Column(name="lvl", type="integer")
-     */
-     private $level;
-
-    /**
-     * @ORM\OneToMany(targetEntity="ForeignUuidCategory", mappedBy="parent")
-     */
-    private $children;
-
-    public function __construct()
-    {
-        $this->id = Uuid\Uuid::uuid1();
-    }
-
-    public function getId()
-    {
-        return $this->id;
-    }
-
-    public function setTitle($title)
-    {
-        $this->title = $title;
-    }
-
-    public function getTitle()
-    {
-        return $this->title;
-    }
-
-    public function setParent(ForeignUuidCategory $parent = null)
-    {
-        $this->parent = $parent;
-    }
-
-    public function getParent()
-    {
-        return $this->parent;
-    }
-
-    public function getRoot()
-    {
-        return $this->root;
-    }
-
-    public function getLeft()
-    {
-        return $this->lft;
-    }
-
-    public function getRight()
-    {
-        return $this->rgt;
-    }
-
-    public function getLevel()
-    {
-        return $this->level;
-    }
-
-    public function getChildren()
-    {
-        return $this->children;
-    }
-
-    public function setChildren($children)
-    {
-        $this->children = $children;
-    }
-
-    /**
-     * @param mixed $root
-     */
-    public function setRoot($root)
-    {
-        $this->root = $root;
-    }
-
-
-}
